+++ credsmash/util.py
@@ -139,8 +139,30 @@ def parse_config(fp):
         config[section] = {}
         for option in cp.options(section):
             config_value = cp.get(section, option)
+            if config_value.startswith("\n"):
+                config_value = _parse_nested(config_value)
             config[section][option] = config_value
     return config
+
+
+def _parse_nested(config_value):
+    # Given a value like this:
+    # \n
+    # foo = bar
+    # bar = baz
+    # We need to parse this into
+    # {'foo': 'bar', 'bar': 'baz}
+    parsed = {}
+    for line in config_value.splitlines():
+        line = line.strip()
+        if not line:
+            continue
+        # The caller will catch ValueError
+        # and raise an appropriate error
+        # if this fails.
+        key, value = line.split('=', 1)
+        parsed[key.strip()] = value.strip()
+    return parsed
 
 
 def set_stream_logger(name='credsmash', level=logging.DEBUG, format_string=None):